requirements.txt
anyio==4.12.0
asn1crypto==1.5.1
bech32==1.2.0
binance-futures-connector==4.1.0
bip_utils==2.10.0
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
coincurve==20.0.0
contourpy==1.3.2
crcmod==1.7
cycler==0.12.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
exceptiongroup==1.3.1
fonttools==4.60.1
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
mplfinance==0.12.10b0
narwhals==2.11.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==12.0.0
plotly==6.4.0
protobuf==6.33.2
py-sr25519-bindings==0.2.3
pycparser==2.23
pycryptodome==3.23.0
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
v4-proto==8.2.0
websocket-client==1.9.0
webcolors==25.10.0
pybit==5.13.0
